pyproject.toml
[tool.poetry]
name = "robox-transfer-poc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.9.1"
webdriver-manager = "^3.8.6"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"
loguru = "^0.7.0"
aiohttp = "^3.8.4"
selenium-wire = "^5.1.0"
pika = "^1.3.2"
asyncpg = "^0.27.0"
nest-asyncio = "^1.5.6"
msgpack = "^1.0.5"
fastapi = "^0.114.0"
sqlalchemy = "^2.0.34"
redis = {extras = ["hiredis"], version = "^5.0.8"}
click = "^8.1.7"
uvicorn = "^0.30.6"
orjson = "^3.10.7"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
